requirements.txt
fastapi==0.85.1
typing==3.7.4.3
pydantic==1.10.2
uvicorn[standard]==0.19.0
sqlalchemy==1.4.42
pymysql==1.0.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
alembic==1.8.1